implement asset system and refactor existing routines to work with it.
rough outline is, project has indexed asset files you can add in-editor; then you can include them in a scene (which then loads the actual asset into memory) and access from inspector

new project -> import asset -> create game object -> attach asset now works with project persistence

+++ engine/src/org/yunghegel/salient/engine/graphics/OutlineDepth.kt
@@ -38,8 +38,8 @@ class OutlineDepth
         val size = 1 / this.size
 
         // float depthMin = ui.outlineDepthMin.getValue() * .001f;
-        val depthMin = depthMin.pow(10.0f) as Float // 0.35f
-        val depthMax = depthMax.pow(10.0f) as Float // 0.9f
+        val depthMin = depthMin.pow(10.0f) // 0.35f
+        val depthMax = depthMax.pow(10.0f) // 0.9f
 
         // TODO use an integer instead and divide w and h
         shader.setUniformf("u_size", Gdx.graphics.width * size, Gdx.graphics.height * size)
